refactor the html and js dashboard functions for re use with future zproc dashbaord
