+++ bin/desi_dailyproc
@@ -8,18 +8,14 @@ No delimiters as it is only a single column (loaded as csv internally)
 
 import sys, os, glob, time, subprocess
 import argparse
-import psutil
 import socket
-import datetime
 import desispec.io
-import fitsio
 from astropy.io import fits
-import numpy as np
 
 ########################
 ### Helper Functions ###
 ########################
-from desispec.desi_proc_dashboard import what_night_is_it, check_running,\
+from desispec.workflow.proc_dashboard_funcs import what_night_is_it, check_running,\
                                          get_file_list, get_skipped_expids
 
 def find_new_nightexp(night, fileglob, known_exposures):
