+++ bin/desi_proc_dashboard
@@ -7,12 +7,12 @@
 """
 This script uses 
 Usage:
-desi_proc_dashboard.py -n all --n_night 60  --output_dir /global/project/projectdirs/desi/www/users/zhangkai/desi_proc_dashboard/ --output_url https://portal.nersc.gov/project/desi/users/zhangkai/desi_proc_dashboard/
+proc_dashboard_funcs.py -n all --n_night 60  --output_dir /global/project/projectdirs/desi/www/users/zhangkai/desi_proc_dashboard/ --output_url https://portal.nersc.gov/project/desi/users/zhangkai/desi_proc_dashboard/
 
 """
 
 import sys
-from desispec.desi_proc_dashboard import parse,main
+from desispec.scripts.procdashboard import parse,main
 
 args = parse(options=None)
 sys.exit(main(args))
